Add Stop class to force downstream tasks to break their loops.
This stops an infinite blocking situation where a task throws an exception, it gets "intercepted" and downstream tasks are scheduled for cancellation but don't _actually_ get cancelled because they're awaiting the queue value from the task that threw the exception.

This puts a value on the queue so the await completes, allowing the event loop to cancel the tasks and the topology to finish.

Needs some cleanup but the basic idea is there.