Trying to get the _right_ exception raised.
This works, but is very brittle. I think it's the right approach though. On receiving a stop or failing, a task should _clear_ the inbound queue(s) before breaking the loop.

Then on gather, if any tasks returned a Fail object, re-raise the exception. A more general form of this might be to clear _all_ queues from top to bottom and raise the exception. Or, maybe clear all queues then pass the failure down the line.