Okay, now we are shutting down all downstream tasks when we get a stop.
This needs to be factored into something that's less ugly, but the idea works.

When a task gets a stop, it forwards the top n times to each output channel, where n is the number of tasks listening to that channel. When each of _those_ tasks gets the message, they forward the message n times to the downstream consumers, where n is the number of tasks.

Note it's possible to dupe messages but that's okay since we aren't waiting for the queues to empty any more, we're waiting for the tasks to complete.